+++ src/Engine/Renderer/Renderer.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Base/Constants.h>
+#include <Renderer/RenderGraph/RenderGraph.h>
 #include <memory>
 
 NAMESPACE_START
@@ -22,12 +23,14 @@ public:
 
     MeshRenderer& getMeshRenderer();
     PostProcessRenderer& getPostProcessRenderer();
+    const RenderGraph::Stats& getLastRenderGraphStats() const { return lastRenderGraphStats_; }
 
 private:
     RenderContext& renderContext_;
     AssetManager& assetManager_;
     std::unique_ptr<MeshRenderer> meshRenderer_;
     std::unique_ptr<PostProcessRenderer> postProcessRenderer_;
+    RenderGraph::Stats lastRenderGraphStats_;
 };
 
 NAMESPACE_END